Fix InputOutputMap tests - not finished yet
