--- engine/test/inputmap/test.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-export LD_LIBRARY_PATH=../../src
-export DYLD_FALLBACK_LIBRARY_PATH=../../src
-./inputmap_test
