+++ engine/test/inputoutputmap/inputoutputmap_test.h
@@ -1,6 +1,6 @@
 /*
   Q Light Controller - Unit test
-  outputmap_test.h
+  inputoutputmap_test.h
 
   Copyright (c) Heikki Junnila
 
@@ -17,14 +17,13 @@
   limitations under the License.
 */
 
-#ifndef OUTPUTMAP_TEST_H
-#define OUTPUTMAP_TEST_H
+#ifndef INPUTOUTPUTMAP_TEST_H
+#define INPUTOUTPUTMAP_TEST_H
 
 #include <QObject>
 
 class Doc;
-
-class OutputMap_Test : public QObject
+class InputOutputMap_Test : public QObject
 {
     Q_OBJECT
 
@@ -33,19 +32,33 @@ private slots:
     void cleanupTestCase();
 
     void initial();
-    void setPatch();
-    void claimReleaseDumpReset();
-    void blackout();
     void pluginNames();
-    void pluginOutputs();
-    void universeNames();
-    void configure();
-    void slotConfigurationChanged();
-    void mapping();
+    void pluginInputs();
+    void configurePlugin();
     void pluginStatus();
+    void profiles();
+    void setPatch();
+    void slotValueChanged();
+    void slotConfigurationChanged();
+    void loadInputProfiles();
+    void inputSourceNames();
+    void profileDirectories();
+
+//    void initial();
+//    void setPatch();
+//    void claimReleaseDumpReset();
+//    void blackout();
+//    void pluginNames();
+//    void pluginOutputs();
+//    void universeNames();
+//    void configure();
+//    void slotConfigurationChanged();
+//    void mapping();
+//    void pluginStatus();
 
 private:
     Doc* m_doc;
 };
 
 #endif
+
