+++ engine/test/test.pro
@@ -16,10 +16,9 @@ SUBDIRS += fixturegroup
 SUBDIRS += function
 SUBDIRS += genericfader
 SUBDIRS += grandmaster
-#SUBDIRS += inputmap
+SUBDIRS += inputoutputmap
 SUBDIRS += inputpatch
 SUBDIRS += mastertimer
-#SUBDIRS += outputmap
 SUBDIRS += outputpatch
 SUBDIRS += qlccapability
 SUBDIRS += qlcchannel
